wheresMyBone/src/wheresmybone/control/GameControl.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package wheresmybone.control;

/**
 *
 * @author Jan
 */
public class GameControl {
    // calculate time left to complete game
    public double calcTimeLeft(double timeLeft, double travelTime) {

  
        if (timeLeft<=0) {
            return -1;
         }
        if (travelTime<0) {
            return -1;
        }

       timeLeft -= travelTime;
       if (timeLeft <= 0) {
             return -1;
       }
       return timeLeft;
       }

    }
